+++ src/driver/periph/crc.h
@@ -2,6 +2,6 @@
 #define __CRC_H__
 
 void crc_init(void);
-float calculate_crc_of_words(uint32_t *data_arr, int size);
+uint32_t calculate_crc_of_words(uint32_t *data_arr, int size);
 
 #endif
